CNN/Handwritten Digit Recognizer/requirements.txt
torch
torchvision
streamlit
numpy
opencv-python
Pillow
scikit-learn